Continue backoffice code quality refactor
