+++ apps/backoffice/src/components/tmdb-reviews/index.tsx
@@ -2,9 +2,6 @@ import type {
   TMDBMatchReview,
   TMDBMatchReviewAction,
   TMDBMatchReviewActionAudit,
-  TMDBMatchReviewReason,
-  TMDBMatchReviewSort,
-  TMDBMatchReviewStatus,
   TMDBReviewCandidate,
 } from '@pop-choice/shared';
 
@@ -19,16 +16,16 @@ import {
   renderStatus,
   StatusBadge,
 } from './reviewPresentation';
+import {
+  buildReviewPageHref,
+  canApplyCandidate,
+  candidateConfidenceGap,
+  getCandidateWarning,
+  isCurrentCandidate,
+  type ReviewFilters,
+} from './helpers';
 
-function ReviewFilterSummary({
-  filters,
-}: {
-  filters: {
-    status: TMDBMatchReviewStatus | 'all';
-    reason: TMDBMatchReviewReason | 'all';
-    sort: TMDBMatchReviewSort;
-  };
-}) {
+function ReviewFilterSummary({ filters }: { filters: ReviewFilters }) {
   return (
     <div className="toolbar-summary" aria-label="Active filters">
       <span className="pill">
@@ -42,39 +39,13 @@ function ReviewFilterSummary({
   );
 }
 
-function buildReviewPageHref({
-  filters,
-  page,
-  pageSize,
-}: {
-  filters: {
-    status: TMDBMatchReviewStatus | 'all';
-    reason: TMDBMatchReviewReason | 'all';
-    sort: TMDBMatchReviewSort;
-  };
-  page: number;
-  pageSize: number;
-}) {
-  const params = new URLSearchParams();
-  params.set('status', filters.status);
-  params.set('reason', filters.reason);
-  params.set('sort', filters.sort);
-  params.set('page', String(page));
-  params.set('pageSize', String(pageSize));
-  return `/tmdb-reviews?${params.toString()}`;
-}
-
 function PaginationControls({
   filters,
   page,
   pageSize,
   totalCount,
 }: {
-  filters: {
-    status: TMDBMatchReviewStatus | 'all';
-    reason: TMDBMatchReviewReason | 'all';
-    sort: TMDBMatchReviewSort;
-  };
+  filters: ReviewFilters;
   page: number;
   pageSize: number;
   totalCount: number;
@@ -131,13 +102,7 @@ function CandidateSummary({ candidates }: { candidates: TMDBReviewCandidate[] })
   if (candidates.length === 0) return <span className="muted">No candidates captured</span>;
 
   const [best, runnerUp] = candidates;
-  const gap =
-    best?.confidence !== null &&
-    best?.confidence !== undefined &&
-    runnerUp?.confidence !== null &&
-    runnerUp?.confidence !== undefined
-      ? best.confidence - runnerUp.confidence
-      : null;
+  const gap = candidateConfidenceGap(candidates);
 
   return (
     <div className="candidate-summary">
@@ -215,11 +180,7 @@ export function ReviewListPage({
   pagination,
 }: {
   reviews: TMDBMatchReview[];
-  filters: {
-    status: TMDBMatchReviewStatus | 'all';
-    reason: TMDBMatchReviewReason | 'all';
-    sort: TMDBMatchReviewSort;
-  };
+  filters: ReviewFilters;
   pagination: {
     page: number;
     pageSize: number;
@@ -305,24 +266,6 @@ export function ReviewListPage({
   );
 }
 
-function getCandidateWarning({
-  review,
-  candidate,
-}: {
-  review: TMDBMatchReview;
-  candidate: TMDBReviewCandidate;
-}): string | null {
-  if (candidate.id === null) return 'Candidate has no TMDB id and cannot be applied.';
-  if (candidate.confidence !== null && candidate.confidence < 0.7) {
-    return 'Low confidence candidate. Verify title, year, and runtime before applying.';
-  }
-  if (candidate.releaseYear !== null && candidate.releaseYear !== review.movieYear) {
-    return `Release year differs from local movie year ${review.movieYear}.`;
-  }
-
-  return null;
-}
-
 function CandidateCard({
   review,
   candidate,
@@ -332,10 +275,9 @@ function CandidateCard({
   candidate: TMDBReviewCandidate;
   index: number;
 }) {
-  const canApply =
-    candidate.id !== null && (review.status === 'open' || review.status === 'deferred');
+  const canApply = canApplyCandidate({ candidate, review });
   const isBest = index === 0;
-  const isCurrent = candidate.id !== null && candidate.id === review.currentMovie?.tmdb_id;
+  const isCurrent = isCurrentCandidate({ candidate, review });
   const warning = getCandidateWarning({ candidate, review });
 
   return (
